Add support for Django 1.7+. Version bump 0.2.4
